EDU-294 refactor: 모듈 정리
